requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
gdown==5.2.0
humanfriendly==10.0
idna==3.10
imageio==2.35.1
imgviz==1.7.5
kiwisolver==1.4.7
labelImg==1.8.6
labelme==5.5.0
lazy_loader==0.4
lxml==5.3.0
matplotlib==3.9.2
mpmath==1.3.0
natsort==8.4.0
networkx==3.3
numpy==2.1.2
onnxruntime==1.19.2
packaging==24.1
pillow==10.4.0
protobuf==5.28.2
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
QtPy==2.4.1
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
soupsieve==2.6
sympy==1.13.3
termcolor==2.4.0
tifffile==2024.9.20
tqdm==4.66.5
urllib3==2.2.3
